Fix broken transaction() function (and standardize other query formatting)
The knex transaction call doesn’t handle the attempted promisification
very well. In the old code it would invoke the Promise’s resolve
function, get the immediate result (`undefined`) and then not do
anything.

This requires all transaction code to manually commit and rollback. The
new way creates a nesting block but properly has the behavior of commit
or rollback based on the success of the callback.

Additionally, standardize formatting of queries within the dao
interface.