Update ESI-related code to use new ESI wrapper
This slightly reorganizes some functions in the old esi.js wrapper (now
eve.js) to fetch a promise for a token and then expose the shared esi
module.  All code seems to function.

Since the new esi module has caching built in, the cache-control table
was removed and the cache logic was removed from the skills code.
However, skills are still persisted to the DB. This can be removed too,
or we can keep that and have error fallbacks on the ESI end point that
read the DB for ‘most up-to-date’ skills available, i.e. if a token has
been revoked.