Add an explicit dependency on a Promise module
Node has native Promise support but it is limited to the bare
specification for Promises. bluebird has more features (like being able
to map with promises instead of composing a Promise.all with a regular
map).

Additionally, the ESI library and knex use bluebird so we are reducing
the number of different Promise libraries loaded into memory.