Implement syncKillboard with ESI (#81)
* Re-arrange JS files to be more consistent with folder structure used in shared/ and front-end/

* Improve README documentation

* Add info columns to citadel table

Include citadel type, if the alliance owns it, and if it is public
access to the alliance. This is synced with what’s on the wiki
currently.

* Add a dao module

The dao module connects to the DB via knexJS and exports functions that
wrap the actual DB queries.

Currently only has very simple accessors and modifiers, and modifiers
are untested (and may need to be updated to have proper transaction
support).

* Make knex module const.

* Add dao as dependency to api module

* Fix path to static resources after server.js move

* Update dao.js to reflect schema changes

* Add ESI dependency

* Normalize transaction usage in schema.js

It’s cleaner to use the trx instance as a builder instead of calling
transacting(trx), especially because the former does not require
manually calling commit and rollback.

Also normalized to using lambdas for the table creation

* Add .idea directory to ignore file

(switched to using WebStorm for IDE)

* Reformat then() newlines to be more consistent with other code.

* Fix broken transaction() function (and standardize other query formatting)

The knex transaction call doesn’t handle the attempted promisification
very well. In the old code it would invoke the Promise’s resolve
function, get the immediate result (`undefined`) and then not do
anything.

This requires all transaction code to manually commit and rollback. The
new way creates a nesting block but properly has the behavior of commit
or rollback based on the success of the callback.

Additionally, standardize formatting of queries within the dao
interface.

* Update dashboard, authentication and skills module to use new transaction function

* Update compiled front-end distribution

* Update ESI-related code to use new ESI wrapper

This slightly reorganizes some functions in the old esi.js wrapper (now
eve.js) to fetch a promise for a token and then expose the shared esi
module.  All code seems to function.

Since the new esi module has caching built in, the cache-control table
was removed and the cache logic was removed from the skills code.
However, skills are still persisted to the DB. This can be removed too,
or we can keep that and have error fallbacks on the ESI end point that
read the DB for ‘most up-to-date’ skills available, i.e. if a token has
been revoked.

* Remove cache-control table

* Move stub output controls to a config parameter

* Update existing JSON stubs to match new data layout

* Move corporation stub into separate file

* Add stub support to skill queue requests

* Address warnings from WebStorm's linter

* Delete old scripts that are no longer necessary

* Remove additional obsolete script

* Add script to manually run a cron task

* Add an explicit dependency on a Promise module

Node has native Promise support but it is limited to the bare
specification for Promises. bluebird has more features (like being able
to map with promises instead of composing a Promise.all with a regular
map).

Additionally, the ESI library and knex use bluebird so we are reducing
the number of different Promise libraries loaded into memory.

* Implement killboard stat sync with ESI API.

* Update esi dependency to use a specific version release

* Update to new ESI wrapper version that has bug fixes