Refactor tooltip to avoid reflows
Before we were reading the width and height of the hosting element
every time -- even if there was no tooltip to show. In pages with
many tooltips (like the roster), this was causing cascading reflow
calculations and was adding appreciable lag. Then I remembered that
flexbox is rad, and found a way to achieve the same effect without
the need for the JS. Woo, flexbox!