Add warning message support to the loading spinner component
Also replaces the PNG version of the warning icon with an SVG version
(color swapped version of error-icon.svg).

The warning message is provided as an external property of the spinner
(since the application determines when a warning is warranted). If the
message is present, then the loading spinner sticks around and shows
the message as text or icon, just like the error message. Error
messages override the warning message.