Missing access tokens report warnings (#113)
Updates the back end to throw an error class when a token is missing. End points that care about this particular exception type catch it and add a warning. These are the skills, skillqueue, queueSummary, and syncKillboard cron. For skills, old data is returned with the warning; for the killboard, the character is skipped; for the skill queue, only the warning is returned. This last case is not significantly different than just allowing the error to be thrown, so perhaps protectedEndpoint.js should be updated instead.

This also updates the front end's loading spinner to handle warning messages that are controlled by its parent. When a warning message is set, the spinner remains visible and displays the message as an icon or text block as configured, using a warning icon instead of the error icon.

This front end infrastructure is largely all that's necessary to support ESI failure hardening as well, but that can now be back end changes that report warnings and the front end will take advantage of it (and currently the routes that use access tokens are the same as those that use ESI).

One thing that this does not do is handle the case when a token refresh fails. At the moment we don't do anything special except allow the exception to propagate but if it's necessary, I can figure out what that type of failure looks like and also have it throw a MissingTokenError.

Fixes #67