Sync zkillboard (#122)
Add (for the time being) a separate task for syncing kill stats via zKillboard. This is blocked by #121 because that modifies the original syncKillboard, and includes a user agent in the configuration.

Once 121 is merged, I will replace the contents of syncKillboard.js with syncZKillboard.js, remove the unnecessary syncZKillboard.js and update the script to use the user-agent defined in the config file.