Logging overhaul (#127)
This PR uses scribe-js to handle logging. It supports logging both to a console and to the file system, which can then be viewed with a web interface. This web viewer has been attached to the admin console in our app and guarded on the server by requiring the session to have read serverLogs privileges, which is assigned only to the admin role.

Fixes #77