Make eve-roster deployable on Dokku (#129)
The biggest change is squashing the front-end and back-end apps down into a single app with all the config files in the root. This was needed to Dokku deployment to be able to pick up the package.json and app.json files and do automatic deployment.

I retained the ability to do hot-loading of webpack modules by using the webpack middleware in place of the webpack dev server. Running the app without NODE_ENV=production will automatically go into this mode.

Other assorted environment variable and config file changes were needed to support the deployment.

TODOs:
* Delete config.local.json and replace with environment variables (managed with node-foreman on dev servers and with dokku in prod) or database records (e.g. the main corp and alt corps for an admin)
* Add a staging instance we can push to for testing our changes on Dokku. Among other things, I discovered that every roster developer is using a case-insensitive file system.
* Add SSL support on the linode and Dokku
* Get a real hostname
* Add some additional deployment users and server admins