Move corp and siggy config into DB
As part of our dokku-ization, we're moving away from file-based
configs. This changes moves the list of "member" corporations as
well as the siggy credentials into the database.

As part of this, it also does the following:
- Creates a special "__admin" role. Accounts with this role have max
  access on all permissions.
- Creates a way to explicitly assign roles to characters.
- Adds a new "quicksetup.js" script that will import setup from
  config.local.json into the DB.
- Rejiggers updatedb script to allow it to be called
  programmatically.

After this change, the process for setting up a dev server will be:
  1. Start server
  2. Log in
  3. Run quicksetup.js script

These steps seem backwards, but the script needs at least one
account to exist so it can assign the "__admin" role to it.