Update display for paused and empty skill queues.  Resolves #145.
Also adds tests on util.time.shortDurationString and changes that to treat
negative durations as absolute value.

It is not an error state for a skill queue to be paused or empty, because we
have no way to tell whether or not a given character's queue *should* be
active without access to XML API account-level data, which may actually be
removed with the deprecation of the XML API and not replaced with equivalent
ESI endpoints.