Simplify roster messaging
Specifically, rely on the corp membership column now returned by the
dao, instead of requiring an extra query to memberCorps.  This change
reverts a lot of the previous changes that had to pass around the
memberCorps map.

Also updates the nomenclature to refer to the messages as alerts
instead of warnings.