Fix error when approving SRPs
The code for setSrpVerdict() was trying to be cute by shoving all
of its work into a single DB query. It didn't work. Now we separate
between info abount the loss (first query) and searching for any
existing reimbursements (later queries, spun off into
findExistingReimbursements().

Also adjusts the payment UI so that "Amount" comes first and is the
button that opens the info window in the client (this was previously
the "Reason" field).