Convert syncRoster to use ESI endpoints
Complete rewrite of syncRoster task to use only the new ESI endpoints.

In order to access membership data, the task must use the access token
of a member character with the Director role. For each corporation, the
task tries to use the credentials of each director until one succeeds
(outside of abnormal operations, the first one should work).