Update ZKillmail interface to more accurately match real world (#362)
Previously we were attempting to distinguish between different kinds of
attackers, but it turns out that this is a fool's errand as the properties
can appear in almost any combination. Better to just treat it as such.

Also marks the "items" section as optional. I've never seen a killmail
without one, but the REST spec says its optional.

Surprisingly, no front-end changes necessary.