Autogenerate battle reports in syncKillmails
Adds an extra step to syncKillmails that generates battle reports.

Battle reports use a relatively simple clustering algorithm. The clusterer
defines a sliding temporal window of 20 minutes. If any of the participants
of a killmail also appear in a existing battle in the window, add that
killmail to the battle. Otherwise create a new battle. If a killmail would
be added to more than one battle, first merge the battles.

Uses a stream pipeline to process killmails. In retrospect, streams are kind
of awful, especially in Typescript. But the backpressure management is
nice in this application, as it allows the system to chew through
an arbitrary number of killmails while using a relatively constant amount
of memory.