Replace ScribeJS with textual logging (#393)
Instead of logging to JSON, we revert to logging to a textual format that's
easy to grep. Moves the actual logging to a separate process (`witness`)
so that output from crashes is still logged correctly.

The `witness` process is now the first process to be started. It then spawns
the actual server as a child process. It reads all stdout/stderr output from
the child and writes it to a rotating text file.