Handle "forbidden" errors from citadel data calls.
Unfortunately, AFAICT there's no way to avoid these errors completely,
as there's no API to determine if a character is on a citadel's ACL
short of just trying to retrieve the citadel info.

So I just have to accept the possibility of getting a 403. When that
happens, I will stop retrieving citadel names for that character.
Hopefully, that will be enough to prevent us from running into ESI's
"no more than 100 errors per minute" limit.