Add support for statically-available Env vars
In general, we _should_ be explicitly injecting our env vars into
the places that they need to go, but a bunch of legacy code is
structured such that the plumbing necessary would be onerous.

Adds the getEnvLegacy() method that any module can call to get access
to the global env vars. Also adds a fake implementation for tests.
Figuring out how to fake out dependencies in Jest was not the most fun.