Defer loading serving code until process has initialized
Otherwise, legacy code that tries to read Env on import will fail because
Env hasn't been initialized yet (should we be doing anything meaningful on
import? No, but that's a story for another day).

Also moves all import statements to the top of the file since that's
how they'll actually be executed now that we're running in EJS module
mode.