Fix "Update SDE" task
Some newer skills have malformed skill requirements, which was causing the
import to be rejected. Sigh. The system is already resilient to such
problems, but added a way for us to allowlist known discrepancies.

Also added a special "dry run" version of the Update SDE task that is only
available on dev builds. Instead of downloading the SDE from the web, it
looks for it in tmp/ and doesn't actually commit any changes to the db.
Useful when you're trying to figure out why the import is failing.