Introduce flows
Implements the `flow` microlibrary, a small system for managing
asynchronous pipeline processing. This is largely a replacement for
using Node Streams in object mode, which were never meant for this
purpose and don't feel great to use in this form.

We should really juse use RxJS instead of this library, but I'm
stubborn and I don't like the shape of its APIs. The structure of
this version is largely inspired by Kotlin flows.

The system supports full backpressure control, so large sources like
files or database queries will not consume excessive memory.