Update SRP sync to use ESI rather than ZKillboard
Implements a new version of syncKillmails that employs flows and
updated market data sources.

Attempts to recreate the metadata provided by zkill as much as possible,
namely the "npc" and fitted/totalValue fields, but we can't recreate
everything, sadly (especially the "solo" field).

Additionally, the ESI endpoint only returns killmails for which a corp
member provided the killing blow, so we will lose access to some kills
that were previously provided by ZKill. however, the SRP system really
only needs access to the losses; the kills exist solely to provide
context for SRP triagers.

A couple of additional pieces of work need to occur after this change.
First, we need to rename all usages of ZKillmail to something more
generic like AnnotatedKillmail. Second, many of our losses do not
appear in ZKill anymore, so we need to provide a detail view page
for every killmail we store.