Implement fetchAverageMarketPrices
Previously we were using fetchJitaSellPrices, which we still maintain,
but honestly we should probably just use the galactic average, which
CCP provides a very efficient and convenient endpoint for.

Adds fetchAverageMarketPrices, which does just that. It caches and
coalesces requests and is just generally quite fast. It stores all ~15k
market prices in memory, but that's probably fine.